fix(io): login-decode parity — lowMemory low-bit mask, GBit width, rev/CRC before RSA (L35-L37)
[N] of the TS-parity fix-all (low tier). Closes L35-L37.

L35 — the login lowMemory flag was decoded with GBool (g1() == 1), but TS
World.ts:2126-2127 reads an info byte and takes the low bit only:
`(info & 0x1) !== 0`. The two diverge for any odd value > 1 (e.g. 3 → TS
true, GBool false). GameLogin now masks the low bit. +pin
TestUnmarshalHeader_LowMemoryMasksLowBit (info 0/1/2/3/255; info=3 is the
decisive case GBool would mis-decode).

L36 — Packet.GBit returned uint8, silently truncating reads wider than 8
bits to their low byte, though the bitmask table spans 32 bits. TS
Packet.gBit (Packet.ts:384) returns a number. Widened the accumulator and
return type to int. Latent today (all callers read <= 7 bits) but a
correctness trap for any future wide bit-read. +pin TestGBitWideValue
(16/12/32-bit round-trips).

L37 — login validation RSA-decrypted before checking revision and CRC. TS
World.ts gates rev (2119) and CRC (2131) on the cleartext header BEFORE
calling rsadec (2139), so a stale-revision or bad-CRC client never burns RSA
CPU — the same rationale as the address rate-limit gate already in
handleLogin. Split GameLogin decode into UnmarshalHeader (opcode, payload
size, revision, lowMemory, archive checksums) and UnmarshalRSA (the
encrypted tail); UnmarshalBinary still calls both for the
encoding.BinaryUnmarshaler contract and round-trip. handleLogin now decodes
the header, validates rev + CRC, then decrypts the RSA block. +pins
TestUnmarshalHeader_DecodesWithoutRSABlock and TestGameLogin_RoundTrip.

+++ pkg/io/packet/packetbit.go
@@ -29,22 +29,26 @@ func (p *Packet) AccessBytes() {
 }
 
 // GBit returns the next n bits in the [Packet].
-func (p *Packet) GBit(n int) uint8 {
+//
+// Returns int to match TS Packet.gBit (Packet.ts:384, returns number): the
+// accumulator must be wider than a byte or reads of more than 8 bits silently
+// truncate. The bitmask table spans 32 bits, so n may be up to 32. L36.
+func (p *Packet) GBit(n int) int {
 	bytePos := p.BitPos >> 3
 	bitsRemaining := 8 - (p.BitPos & 0x7)
-	value := uint8(0)
+	value := 0
 	p.BitPos += n
 
 	for ; n > bitsRemaining; bitsRemaining = 8 {
-		value += (p.Data[bytePos] & uint8(bitmask[bitsRemaining])) << (n - bitsRemaining)
+		value += int(p.Data[bytePos]&uint8(bitmask[bitsRemaining])) << (n - bitsRemaining)
 		bytePos++
 		n -= bitsRemaining
 	}
 
 	if n == bitsRemaining {
-		value += p.Data[bytePos] & uint8(bitmask[bitsRemaining])
+		value += int(p.Data[bytePos] & uint8(bitmask[bitsRemaining]))
 	} else {
-		value += (p.Data[bytePos] >> (bitsRemaining - n)) & uint8(bitmask[n])
+		value += int((p.Data[bytePos] >> (bitsRemaining - n)) & uint8(bitmask[n]))
 	}
 
 	return value
